fload - working primitive
quit - quits fload

a tilde marks bad words in fload, rather
than the question mark.  That's the best
way to tell if fload is currently looping
or not.

The fload loop is pretty much the same as
the quit loop.

Might be named flquit right now.  Have forgotten. ;)

	modified:   Cortex-Forth.ino

On branch fload-bb-aa