working code - very good indeed.
The program now minimally reads a single
integer from the flashROM file, and
sucessfully loads it onto the stack.

It also preloads a zero onto the stack,
first, which is an error of unknown
origin.

Changes that make this happen follow, in
successive commits after this one. ;)

	modified:   dog.cpp

On branch fload-bb-ee-aa