rolled back some changes by preserving them in the deeper repository history.
Did a reset; made a copy; restored effects of
reset and then copied over this file Cortex-Forth.ino
with an earlier revision.

Not sure that matches anything exactly, as some
edits may've been introduced to make this variant
unique.

Results: with older one-liner test.fs program (stack
elements only, no compiling) the present variant
seems to behave well.

Evident: the _COLON word calls the _PARSE word, which
reset tib to the empty string.  This is in conflict
with how files are read in and compiled.

	modified:   Cortex-Forth.ino

On branch fload-bb-ee-aa