used as master
NEW BRANCH

This branch, going forward, is used much like
a branch named 'master' would be.  At least that
was the thought behind it.  Unsure what the
consequences of such a decision, are. ;)

It's good enough and stable enough right now
to deploy (carefully) in limited applications,
unless a new bug is noticed.

 $ date
Mon Aug 12 00:39:22 UTC 2019

 $ git branch | cat | egrep side
* side-fload-bb-ee-aa-aa

 $ git checkout -b good-compiler-aa
Switched to a new branch 'good-compiler-aa'

	modified:   _branch_storyline

On branch good-compiler-aa