Hardware compatibility
A series of conditionals for the C preprocessor,
which makes decisions based on hardware available.

Initially this is to detect which Adafruit target
board is present, and to use that information
to enable/disable dotStar (RGB LED) support.

ItsyBitsyM4 is the only supported target during
this iteration. ;)

	new file:   compatibility.h

On branch good-compiler-aa