edits
The sam editor is in pretty good shape for
a primitive demo - it will modify each location
in an 'allot'ed buffer, in sequence, and skip
over the first byte as well as the last two
bytes alloted.

The first byte seems to be used for something;
The second to last (or the last) byte may be
set to the null.  Leaving room for that, plus
one extra byte (don't remember why, at this
moment).

	modified:   flash_ops.cpp

On branch non-usart-b-31_aug-aa-