Everthing above : sam ; is good.
Remember, the kludge is to extend lines
without line breaks, when separate print
statements are used.

Examine the output file stored on flashROM
to see how this functions - the code is fairly
well formed as stored on file, in that each
colon definition is unbroken by line breaks.

Maybe - unverified. ;)

	modified:   sam.cpp