comment escape processing added
The program will now scan for 0x0d after it sees
a leading \ (0x92) char.

It does seem to need two leading spaces at BOL
(beginning of line).

(See just above the definition for 'hix' which is
(itself) a demonstration word that will go away, soon).

No provision for wanting to use 0x92 anywhere in any
program; the expectation is that your program source
code will be (remakably) free of 0x92 when not used
to source a comment. ;)

	modified:   Cortex-Forth.ino

On branch non-usart--testing-a