Comment processing, formatting
Also includes a demo on the namespace structure,
showing that abca abcb abcc are equivalent
definition names for colon definitions.
abcx which does not exist is also a valid
synonym for the last definition that uses
this name structure (and so is abcq abcp &c.)

In short, word name lengths must differ, when
the first three letters naming the word are
identical.

The program only looks at the first three
characters that name the word, and the length
of the name of the word, but not at the content
of that name beyond the first three characters.

	modified:   src/sam.cpp

On branch non-usart--testing-a