prior release of this program
Current documentation effort is centered on this
exact release of the program (and so, most fully
obtains in relation to it, over other versions).

The very next commit is the reference version, which
is meant to be functionally identical to this current
commit, except for how the files are packed into a
practical archive (.ZIP vs .tar.gz) and both are from
the github Release page for the repository, same exact
version.

	new file:   Cortex-Forth-0.2.0-alpha.3.zip

On branch non-usart--flash_bug-aa-